requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
cloudinary==1.28.1
colorama==0.4.4
dnspython==2.2.0
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==9.0.0
pixelsort==1.0.1
pymongo==4.0.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.8
Werkzeug==2.0.2
gunicorn==20.1.0
